refactor: moved geo scale logic to separate function
